+++ test/functional/wiki_controller_test.rb
@@ -65,52 +65,6 @@ class WikiControllerTest < ActionController::TestCase
                   text: '@alphanodes'
   end
 
-  def test_show_user_with_id
-    @request.session[:user_id] = 1
-    @page.content.text = 'user#1'
-    @page.content.save!
-    get :show, project_id: 1, id: @page_name
-    assert_response :success
-    assert_template 'show'
-    assert_select 'a[href=?]', '/users/1',
-                  text: 'admin'
-  end
-
-  def test_show_user_with_id_fullname
-    @request.session[:user_id] = 1
-    @page.content.text = 'user[fl]#1'
-    @page.content.save!
-    get :show, project_id: 1, id: @page_name
-    assert_response :success
-    assert_template 'show'
-    assert_select 'a.user', text: 'Redmine Admin'
-    assert_select 'a[href=?]', '/users/1',
-                  text: 'Redmine Admin'
-  end
-
-  def test_show_user_with_name
-    @request.session[:user_id] = 2
-    @page.content.text = 'user:jsmith'
-    @page.content.save!
-    get :show, project_id: 1, id: @page_name
-    assert_response :success
-    assert_template 'show'
-    assert_select 'a[href=?]', '/users/2',
-                  text: 'jsmith'
-  end
-
-  def test_show_user_with_name_fullname
-    @request.session[:user_id] = 2
-    @page.content.text = 'user[fl]:jsmith'
-    @page.content.save!
-    get :show, project_id: 1, id: @page_name
-    assert_response :success
-    assert_template 'show'
-    assert_select 'a.user', text: 'John Smith'
-    assert_select 'a[href=?]', '/users/2',
-                  text: 'John Smith'
-  end
-
   def test_show_last_updated_by_marco
     @request.session[:user_id] = 1
     @page.content.text = '{{last_updated_by}}'